comments resovled

+++ market_simulation/models/stock_data.py
@@ -4,10 +4,10 @@ from .base import BaseModel
 
 class StockData(BaseModel):
     ticker = models.CharField(max_length=30)
-    open_price = models.IntegerField()
-    close_price = models.IntegerField()
-    high = models.IntegerField()
-    low = models.IntegerField()
+    open_price = models.FloatField()
+    close_price = models.FloatField()
+    high = models.FloatField()
+    low = models.FloatField()
     volume = models.IntegerField()
 
     class Meta:
